tests/wpt/web-platform-tests/workers/WorkerGlobalScope_ErrorEvent_message.htm

<!DOCTYPE html>
<title> WorkerGlobalScope onerror event handler argument: message </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var message = 'Error Message';
  var worker = new Worker('./support/ErrorEvent.js');
  worker.onmessage = t.step_func_done(function(e) {
    assert_greater_than(e.data.message.indexOf(message), -1);
  });
  worker.postMessage(message);
});
</script>
